refactor(api): move joiners in dataProviders
